python-stats-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
espn-api==0.30.0
requests==2.31.0
gunicorn==21.2.0
numpy==1.26.4
psycopg2-binary==2.9.9